uint64_t

+++ 01_software/sim/circuits/shifters.c
@@ -1,7 +1,7 @@
 #include "shifters.h"
 
-bus64 shift_left(bus64 a, unsigned int n) {
-  for (unsigned int j = 0 ; j < n; ++j) {
+bus64 shift_left(bus64 a, uint64_t n) {
+  for (uint64_t j = 0 ; j < n; ++j) {
     for (int i = BUS64_WIDTH - 1; i > 0; --i) {
       bus64_setbit(&a, i, bus64_getbit(a, i - 1));
     }
@@ -10,8 +10,8 @@ bus64 shift_left(bus64 a, unsigned int n) {
   return a;
 }
 
-bus64 shift_right(bus64 a, unsigned int n) {
-  for (unsigned int j = 0; j < n; ++j) {
+bus64 shift_right(bus64 a, uint64_t n) {
+  for (uint64_t j = 0; j < n; ++j) {
     for (int i = 0; i < BUS64_WIDTH - 1; ++i) {
       bus64_setbit(&a, i, bus64_getbit(a, i + 1));
     }
